requirements.txt
django-extensions
South
django-crispy-forms
django-debug-toolbar
django-xadmin
django-jsonview
django-model-utils
DjangoUeditor
